Add saved payment method selection to checkout page
Implement functionality to display and select saved payment methods during checkout. Update the checkout template to show a list of saved cards with an option to use a new card. Add a toggle to show/hide new card input fields and include a checkbox to save new card details for future use. Corresponding updates in views and models support this feature.